pyproject.toml
[project]
name = "python-tjbotlib"
version = "0.1.0"
description = "Python library for writing TJBot recipes"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "rpi-lgpio>=0.6; sys_platform == 'linux'",
    "pigpio>=1.78; sys_platform == 'linux'",
    "rpi-ws281x>=5.0.0; sys_platform == 'linux'",
    "gpiozero>=2.0",
    "pyalsaaudio>=0.10.0; sys_platform == 'linux'",
    "ibm-watson>=7.0.0",
    "azure-cognitiveservices-speech>=1.34.0",
    "google-cloud-speech>=2.0.0",
    "sherpa-onnx>=1.10.0",
    "tomli>=2.0.1; python_version < '3.11'",
    "webcolors>=1.13",
    "spidev>=3.6; sys_platform == 'linux'",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
    "InquirerPy>=0.3.4",
]

[tool.hatch.build.targets.wheel]
packages = ["src/tjbot"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
